backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
scikit-learn==1.5.2
plotly==5.24.1
reportlab==4.2.2
python-dotenv==1.0.1
kaleido==0.2.1
httpx>=0.25.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
sentence-transformers>=2.2.2